Move API C++ files to include/tscpp
