+++ plugins/experimental/nexthop_strategy_consistenthash/plugin_consistenthash.cc
@@ -41,8 +41,8 @@
 #include "tscore/ink_assert.h"
 #include "ts/ts.h"
 #include "ts/remap.h"
-#include "ts/nexthop.h"
-#include "ts/parentresult.h"
+#include "tscpp/api/nexthop.h"
+#include "tscpp/api/parentresult.h"
 
 void
 TSRemapDeleteInstance(void *ih)
